+++ docs/source/history.rst
@@ -2,6 +2,7 @@ Release history
 ===============
 
 Version 0.6
+  - Log a warning when the sink-input enters the ``suspended`` state.
   - Fix assertion error upon ``exit`` pulseaudio event.
   - Support PipeWire. No change is needed to support PipeWire. The test suite
     runs successfully on PipeWire.